+++ source/crisscross/avltree.h
@@ -51,7 +51,7 @@ namespace CrissCross
 				AVLNode<Key, Data, OwnsKeys> *m_root;
 
 				/*! \brief The current tree size. */
-				size_t m_size;
+				uint32_t m_size;
 
 				/*! \brief Result of tree operation */
 				enum {
@@ -250,7 +250,7 @@ namespace CrissCross
 				/*!
 				 * \return Size of the tree.
 				 */
-				inline size_t size() const
+				inline uint32_t size() const
 				{
 					return m_size;
 				}
@@ -282,7 +282,7 @@ namespace CrissCross
 				/*!
 				 * \return Memory usage in bytes.
 				 */
-				size_t mem_usage() const;
+				uint32_t mem_usage() const;
 
 #if !defined(DISABLE_DEPRECATED_CODE)
 				/*
